% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PrismReport.R
\name{summary.PrismReport}
\alias{summary.PrismReport}
\title{Summarize a PrismReport}
\usage{
\method{summary}{PrismReport}(object, ...)
}
\arguments{
\item{object}{A \code{PrismReport} object returned by \code{\link{prism}}.}

\item{...}{Additional arguments passed to summary methods.}
}
\value{
Invisibly returns the input \code{PrismReport} object.
}
\description{
Summarize a PrismReport
}
\examples{
report <- prism(airquality, target = "Ozone")
summary(report)

}
