% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PrismReport.R
\name{plot.PrismReport}
\alias{plot.PrismReport}
\title{Plot a PrismReport}
\usage{
\method{plot}{PrismReport}(x, ...)
}
\arguments{
\item{x}{A PrismReport object.}

\item{...}{Additional arguments.}
}
\description{
Plot a PrismReport
}
