% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PrismReport.R
\name{print.PrismReport}
\alias{print.PrismReport}
\title{Print a PrismReport}
\usage{
\method{print}{PrismReport}(x, ...)
}
\arguments{
\item{x}{A \code{PrismReport} object returned by \code{\link{prism}}.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
Invisibly returns the input \code{PrismReport} object.
}
\description{
Print a PrismReport
}
\examples{
report <- prism(airquality, target = "Ozone")
print(report)

}
