% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PrismReport.R
\name{summary.PrismReport}
\alias{summary.PrismReport}
\title{Summarize a PrismReport}
\usage{
\method{summary}{PrismReport}(x, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{A PrismReport object.}
}
\description{
Summarize a PrismReport
}
