% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PrismReport.R
\name{summary.PrismReport}
\alias{summary.PrismReport}
\title{Summarize a PrismReport}
\usage{
\method{summary}{PrismReport}(object, ...)
}
\arguments{
\item{object}{A PrismReport object.}

\item{...}{Additional arguments.}
}
\description{
Summarize a PrismReport
}
\examples{
report <- prism(airquality, target = "Ozone")
summary(report)

}
